fix "transparency" of floating tables in docx import (fdo#77723)
Normal tables do not have any fill background on their own, i.e. it's
possible to see e.g. an image in the background, so if we put a table
inside a frame in order to make it floating, make the frame transparent
in order to achieve the same result.

Change-Id: I77b1c8f847af5355c4238bfa089d760dd6121a69