Revert "More --disable-database-connectivity work"
Nah, this was not a good idea. We do need to keep the db related
fields around in some way for document roundtripping. Need to do
database functionality disabling at a different level. If at all, I am
starting to lose trust that it can meaningfully and cleanly be done.

This reverts commit cfaed8d23a1142cc635fd6affca30f5894199ef6.