ooxml: fix cell theme color and shade preservation
The cell property w:shd was not being properly preserved when the
shade value was different from "clear". The shading affects the cell
color and the exporter was not taking that into account when comparing
the cell color with the original color to know if the user had changed
it.

Besides, we were not preserving the attributes themeColor, themeTint
and themeShade.

I have modified the existing unit test testTableThemePreservation to
add a check for those new attributes.

Change-Id: I06d2e668486803cba039eacb717a69413bd5a1df