fdo#76261 Get rid of unneeded and malfunctioning maItemListRect.
maItemListRect serves as a sort of bounding box for what items /could/
be visible on screen, however it is only used for determining the id
of an item we have just clicked on -- given that we already know that
we are clicking inside the visible area this is an unnecessary check.

In fact this check is erronous as we no longer limit ourselves to "whole"
rows as of 7a7b2124fc86afce4e1b853c7500821a3a9010b9 -- i.e. the previous
assumption was that a view large enough for n rows will show precisely
n (whole) rows, whereas we might be showing (n-1) complete rows, and a
fraction of a row at the top and bottom of the screen, i.e. a total of
n+1 rows, where maItemRect only encloses the top n rows, meaning we
were erronously ignoring clicks on the (partly visible) lowest n+1 row.

Change-Id: Ia52ed2e0d217a9f077cef86ee9c1847252844995