fdo#75519: fix chapter cross references
It would be possible to modify the condition in
SwChapterField::ChangeExpansion() to check for rTxtNd.GetNumRule() as
well and make the bugdoc work again, but do we really want to use
"chapter reference" to refer to plain numbering?  Perhaps it would be
better to add a "numbering reference" field type; without investigating
what Word actually does here and what the actual problem is that was
supposedly fixed here (which is not possible given there's no reproducer
document) it's best for now to just revert it.

(regression from commit a7ce680903c0d6aedd430c9a62e3c9ba2fbfcae1)

Change-Id: I7eb9ea46e42729c53c350ff450d49a1eb84ee8eb