Consolidate CreateTempName_Impl and lcl_createName more faithfully
This is a follow-up to 18f5121d3f6926c9eafac1429cb5959b3f25adba "Consolidate
CreateTempName_Impl and lcl_createName," which apparently lead to hard-to-
diagnose sporadic "make check" failures on some machines.  Hope is that bLock
now more carefully tracking the behavior of the original two functions fixes
that.

Change-Id: I37ba00545f5921fd603ff9f1ea46927c42e88899