fdo#78149: set the ViewPort even when painting to a MetaFile
ObjectContactOfPageView::DoProcessDisplay() has a special case for
painting to MetaFiles, and does not set the ViewPort from the passed-in
RedrawArea unless it's a printing or PDF export, with the result
that for a plain MetaFile all fly-frames in a Writer document
are painted, whether visible or not.

Since f9c0624c55b2d3ecdd46754dfd6db586ef14fa6c now calls
GetPreviewMetaFile() after every document load, this issue is more
visible.

Since there is no obvious reason to ever ignore a passed RedrawArea,
simply always use it for the ViewPort, not just when printing.

Change-Id: I256710f1476181f567069b45a2a494b0d2064d6b