ODS load perf: accelerate CompileXML for formula by stringify and compare.
For long columns of repeated formula, it is far faster to convert a formula
to a string and compare to see if the next cell in a column matches, than
it is to parse the formula string itself (sad but true).

This saves 29s of 131s of load time for my (admittedly large) ODS file.

Change-Id: I48b613eb7131d6eb3902695aa30a1aa1a9ea5f6a