reportdesign: Fix applying the last used color
The last used color is stored inside
SvxColorToolBoxControl, so we must call its Select method
to do that. (This also fixes using the last shape, See
commit ce842113ae728e8995c7c5398204a20d0707e843 for a
fix of exactly the same problem in chart2.)

For the font color, there was also another problem:
The SID used for that button is SID_ATTR_CHAR_COLOR2.
The problem is that SvxColorToolBoxControl::Select
actually calls for that SID the ".uno:CharColorExt"
command instead of ".uno:FontColor" (which activates
in writer the format paintbrush mode). Given that since
commit d6802a74ae1765df66f8524b962909dac92cc9a6 this
command also receives the last used color as argument,
the easiest solution is to register ".uno:CharColorExt"
here as well.

Change-Id: I2db25a7fb537ce10dfec7b1c2d049dd77d9f8f6b