Don't forget to dispose the component here.
Otherwise, none of the associated SfxViewFrame, SfxBaseController,
SfxObjectShell, SfxMedium etc would get destroyed which would ultimately
lead to the .~lock.foo file not being deleted upon exit.

But for this to work, the client code has to make sure that its
LibreOfficeDocument instance gets destroyed when it finishes its thing.

Change-Id: Id45eda526c1496712e4314a947e994eba5c3fa68