Revert "Add function to catenate n copies of a string"
COMPHELPER_DLLPUBLIC of an inline function is wrong. Nevermind that I just
followed the example of the preceding functions.

And it is too bikeshedding-inducing anyway to use recursion.

This wasn't used anywahere yet anyway in committed code, just locally in a
tree of mine. Will add a local copy there.

This reverts commit 9a495c9f143c5c23354280fa65e13982d86a2903.