gbuild: tweak var2file so it handles some shell metacharacters
Always quote the passed string; this cannot handle arbitrary strings
like $(file) can, but everything that is needed currently.

Also, try to ensure that output does not start with a space.

Note: there needs to be a newline at the end, otherwise helpex
ignores the last item on the line.

Change-Id: I51a4058591cc4f12dd878c2d113bd5cfc8c22d61