2nd part of bnc#870233: import font color from color fragment for SmartArts
SmartArt import ignores some fragments during import if
drawing fragment exists, which seems to be not complete.
In this case font style is blank (white) in data (and drawing)
fragment and the real value is defined in the ignored color fragment.

So first make color fragment parsing work, then apply font
color of "node0" style on nodes of the SmartArt.

Actually, it's a workaround, because "node0" style label
is hardcoded, for a proper solution layout fragment should
be parsed too to get the right style label, but
it interferes with the drawing fragment by now.

Change-Id: I7db89176a07eee928563d42d3896fbd02190dfa8