SwDoc::DelLayoutFmt: don't delete content of a draw format
In case a shape (has a draw format) has a textbox (RES_CNTNT of the draw
format), then that's just a pointer to that content, but the draw format
doesn't own it: the matching fly format does. So ignore that content
when deleting the layout format in case of draw formats: that ensures
when both the draw and the fly format is deleted, deletion is only
performed once.

Change-Id: Idb4bb19130a6b9acd8f8d3710b9982801b416dda